requirements.txt
# Class-SNA v2.0 Dependencies
# Flask Web Framework

# Flask 핵심
Flask==3.0.0
Flask-Session==0.5.0
Flask-Cors==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# 데이터 처리
pandas==2.1.4
numpy==1.26.3
openpyxl==3.1.2
xlsxwriter==3.1.9

# 네트워크 분석
networkx==3.2.1
python-louvain==0.16
scipy==1.12.0

# 시각화
matplotlib==3.8.2
plotly==5.18.0
pyvis==0.3.2
kaleido==0.2.1

# Google 연동
gspread==6.0.0
google-auth==2.27.0
google-generativeai==0.4.0

# 유틸리티
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
urllib3==2.2.0
certifi==2024.2.2

# 세션/캐시 (Redis)
redis==5.0.1

# 데이터베이스 (PostgreSQL)
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# 인증 (OAuth 2.0)
Flask-Login==0.6.3
Authlib==1.3.0

# 폰트 지원
fonttools==4.46.0
matplotlib-inline==0.1.6

# 보안
python-dotenv==1.0.0
